==> 4_merge_variable.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from datetime import datetime
from sklearn.preprocessing import StandardScaler
import math

import warnings
warnings.filterwarnings("ignore")

off_basic = pd.read_pickle("../new_data/offline/basic.pkl")
off_time = pd.read_pickle("../new_data/offline/time.pkl")
off_group = pd.read_pickle("../new_data/offline/off_group.pkl")

on_basic = pd.read_pickle("../new_data/online/basic.pkl")
on_time = pd.read_pickle("../new_data/online/time.pkl")
online = pd.read_pickle("../new_data/online/online.pkl")
on_group = pd.read_pickle("../new_data/online/ongroup.pkl")

def transform_quantile(df,quantile_set):
    df_percent = np.percentile(df,quantile_set)
    df_quant = list(pd.unique(df_percent))
    df_quant.insert(0,-1)
    label = range(len(df_quant)-1)
    df_rank = pd.cut(df,df_quant,labels=label)
    df_rank = df_rank.astype(int)
    return(df_rank)
quantile_set = [(i+1)*5 for i in range(20)]

off_id = off_basic.copy(deep=True)
off_id['count'] = 1
off_id = pd.DataFrame(off_id['count'])

on_id = on_basic.copy(deep=True)
on_id['count'] = 1
on_id = pd.DataFrame(on_id['count'])

# Basic 

online_basic = pd.merge(on_basic,off_id,on='clnt_id',how='outer')
online_basic = online_basic.fillna(0)
online_basic = online_basic.drop('count',axis=1)

offline_basic = pd.merge(off_basic,on_id,on='clnt_id',how='outer')
offline_basic = offline_basic.fillna(0)
offline_basic = offline_basic.drop('count',axis=1)

online_basic = online_basic.apply(lambda x: transform_quantile(x,quantile_set))
offline_basic = offline_basic.apply(lambda x: transform_quantile(x,quantile_set))

online_basic.to_pickle("../new_data/variables/online_basic.pkl")
offline_basic.to_pickle("../new_data/variables/offline_basic.pkl")

plt.figure(figsize=(20,10))
plt.subplot(2,4,1)
plt.title('buy_ct')
plt.plot(sorted(on_basic['buy_ct']))
plt.subplot(2,4,2)
plt.title('buy_am')
plt.plot(sorted(on_basic['buy_am']))
plt.subplot(2,4,3)
plt.title('tran_freq')
plt.plot(sorted(on_basic['tran_freq']))
plt.subplot(2,4,4)
plt.title('tran_ct')
plt.plot(sorted(on_basic['tran_ct']))
plt.subplot(2,4,5)
plt.title('tran_am')
plt.plot(sorted(on_basic['tran_am']))
plt.subplot(2,4,6)
plt.title('clac_var')
plt.plot(sorted(on_basic['clac_var']))
plt.subplot(2,4,7)
plt.title('clac_ct')
plt.plot(sorted(on_basic['clac_ct']))
plt.subplot(2,4,8)
plt.title('clac_am')
plt.plot(sorted(on_basic['clac_am']))
plt.show()

plt.figure(figsize=(20,10))
plt.subplot(2,4,1)
plt.title('buy_ct')
plt.plot(sorted(online_basic['buy_ct']))
plt.subplot(2,4,2)
plt.title('buy_am')
plt.plot(sorted(online_basic['buy_am']))
plt.subplot(2,4,3)
plt.title('tran_freq')
plt.plot(sorted(online_basic['tran_freq']))
plt.subplot(2,4,4)
plt.title('tran_ct')
plt.plot(sorted(online_basic['tran_ct']))
plt.subplot(2,4,5)
plt.title('tran_am')
plt.plot(sorted(online_basic['tran_am']))
plt.subplot(2,4,6)
plt.title('clac_var')
plt.plot(sorted(online_basic['clac_var']))
plt.subplot(2,4,7)
plt.title('clac_ct')
plt.plot(sorted(online_basic['clac_ct']))
plt.subplot(2,4,8)
plt.title('clac_am')
plt.plot(sorted(online_basic['clac_am']))
plt.show()

# time 

online_time = pd.merge(on_time,off_id,on='clnt_id',how='outer')
online_time = online_time.fillna(0)
online_time = online_time.drop('count',axis=1)

offline_time = pd.merge(off_time,on_id,on='clnt_id',how='outer')
offline_time = offline_time.fillna(0)
offline_time = offline_time.drop('count',axis=1)

online_time = online_time.apply(lambda x: transform_quantile(x,quantile_set))
offline_time = offline_time.apply(lambda x: transform_quantile(x,quantile_set))

online_time.to_pickle("../new_data/variables/online_time.pkl")
offline_time.to_pickle("../new_data/variables/offline_time.pkl")

# group 

online_group = pd.merge(on_group,off_id,on='clnt_id',how='outer')
online_group = online_group.fillna(0)
online_group = online_group.drop('count',axis=1)

offline_group = pd.merge(off_group,on_id,on='clnt_id',how='outer')
offline_group = offline_group.fillna(0)
offline_group = offline_group.drop('count',axis=1)

online_group = online_group.apply(lambda x: transform_quantile(x,quantile_set))
offline_group = offline_group.apply(lambda x: transform_quantile(x,quantile_set))

online_group.to_pickle("../new_data/variables/online_group.pkl")
offline_group.to_pickle("../new_data/variables/offline_group.pkl")

# online 

online = online.apply(lambda x: transform_quantile(x,quantile_set))

online.to_pickle("../new_data/variables/online.pkl")
