==> 6_train_test.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from datetime import datetime
from sklearn.preprocessing import StandardScaler
import math

import warnings
warnings.filterwarnings("ignore")

transaction = pd.read_pickle("../new_data/data_set/transaction.pkl")

# Online 

train = transaction.loc[transaction['date_time'] < datetime.strptime('2019-09-15','%Y-%m-%d')]
train = train.loc[train['biz_unit']=='A']
item_list = [str(i).zfill(4) for i in range(1,1668)]
train = train.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
train.to_pickle("../new_data/train_test/online_train.pkl")

test = transaction.loc[transaction['date_time'] >= datetime.strptime('2019-09-15','%Y-%m-%d')]
test = test.loc[test['biz_unit']=='A']
test = test.sort_values(['clnt_id','date_time'])
test_id = test[['clnt_id','trans_id']]
test_id = test_id.drop_duplicates(keep='first')
test_id = test_id.drop_duplicates('clnt_id',keep='first')
test = pd.merge(test_id,test,on=['clnt_id','trans_id'],how='left')
test = test.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
test = test.applymap(lambda x : 1 if x >= 1 else 0)
test.to_pickle("../new_data/train_test/online_test.pkl")

# Offline 

train = transaction.loc[transaction['date_time'] < datetime.strptime('2019-09-15','%Y-%m-%d')]
train = train.loc[train['biz_unit']=='B']
item_list = [str(i).zfill(4) for i in range(1,1668)]
train = train.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
train.to_pickle("../new_data/train_test/offline_train.pkl")

test = transaction.loc[transaction['date_time'] >= datetime.strptime('2019-09-15','%Y-%m-%d')]
test = test.loc[test['biz_unit']=='B']
test = test.sort_values(['clnt_id','date_time'])
test_id = test[['clnt_id','trans_id']]
test_id = test_id.drop_duplicates(keep='first')
test_id = test_id.drop_duplicates('clnt_id',keep='first')
test = pd.merge(test_id,test,on=['clnt_id','trans_id'],how='left')
test = test.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
test = test.applymap(lambda x : 1 if x >= 1 else 0)
test.to_pickle("../new_data/train_test/offline_test.pkl")

# Total 

train = transaction.loc[transaction['date_time'] < datetime.strptime('2019-09-15','%Y-%m-%d')]
item_list = [str(i).zfill(4) for i in range(1,1668)]
train = train.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
train.to_pickle("../new_data/train_test/train.pkl")

test = transaction.loc[transaction['date_time'] >= datetime.strptime('2019-09-15','%Y-%m-%d')]
test = test.sort_values(['clnt_id','date_time'])
test_id = test[['clnt_id','trans_id']]
test_id = test_id.drop_duplicates(keep='first')
test_id = test_id.drop_duplicates('clnt_id',keep='first')
test = pd.merge(test_id,test,on=['clnt_id','trans_id'],how='left')
test = test.groupby(['clnt_id', 'pd_c'])['buy_ct'].count().unstack().reindex(item_list, axis=1).fillna(0)
test = test.applymap(lambda x : 1 if x >= 1 else 0)
test.to_pickle("../new_data/train_test/test.pkl")
